requirements.txt
oracledb>=2.0
pyodbc>=5.1
PyYAML>=6.0